Reorganize time variables.
