+++ reanimate.cabal
@@ -93,6 +93,10 @@ library
                       Reanimate.Driver.Compile
                       Reanimate.Misc
                       Paths_reanimate
+                      Reanimate.Scene.Core
+                      Reanimate.Scene.Var
+                      Reanimate.Scene.Sprite
+                      Reanimate.Scene.Object
   build-depends:
     base                 >=4.10 && <5,
     JuicyPixels          >=3.3.3,
